+++ include/GafferScene/TypeIds.h
@@ -155,6 +155,7 @@ enum TypeId
 	LocaliseAttributesTypeId = 110610,
 	PrimitiveSamplerTypeId = 110611,
 	ClosestPointSamplerTypeId = 110612,
+	CurveSamplerTypeId = 110613,
 
 	PreviewGeometryTypeId = 110648,
 	PreviewProceduralTypeId = 110649,