+++ python/GafferSceneTest/__init__.py
@@ -148,6 +148,7 @@ from ShufflePrimitiveVariablesTest import ShufflePrimitiveVariablesTest
 from EditScopeAlgoTest import EditScopeAlgoTest
 from LocaliseAttributesTest import LocaliseAttributesTest
 from ClosestPointSamplerTest import ClosestPointSamplerTest
+from CurveSamplerTest import CurveSamplerTest
 
 from IECoreScenePreviewTest import *
 from IECoreGLPreviewTest import *